# perf: non-recursive implementation of euclidian gcd in balanced pool